+++ src/pages/ManagePage/Artist/ArtistList.tsx
@@ -3,7 +3,7 @@ import { useRef, useEffect, useState } from 'react';
 
 import { useGetArtistsQuery } from '../../../features/artists/services/artistsApiSlice';
 import { ArtistContent } from '../../../types/artistsType';
-import { ArtistListSection } from '../style';
+import { ArtistListSection } from '../ManagePageStyle';
 
 import ArtistListItem from './ArtistListItem';
