+++ src/router.tsx
@@ -9,8 +9,8 @@ import EditEvent from './pages/EditEvent/EditEvent';
 import EditReview from './pages/EditReview/EditReview';
 import EventList from './pages/EventList/EventList';
 import Main from './pages/Main/Main';
-import ManagePage from './pages/ManagePage';
-import MyPage from './pages/MyPage';
+import ManagePage from './pages/ManagePage/ManagePage';
+import MyPage from './pages/MyPage/MyPage';
 import NotFound from './pages/NotFound/NotFound';
 import ResetPassword from './pages/ResetPassword/ResetPassword';
 import ReviewDetail from './pages/ReviewDetail/ReviewDetail';
